move MLP to 'layers'
